common: Redo component selection logic
Maintainers should not need to manually select
the appropriate variant of a component for
their SoC. We know what does and does not
work for an SoC. Automatically select
the necessary component. This allows us
to be more flexible when new SoCs launch
and components may need to be renamed or
refactored.

Change-Id: Ibe9815ba7f70ec760c7b0cbd3771707af2e59854
Signed-off-by: chrisl7 <wandersonrodriguesf1@gmail.com>